refactor: move vtex & newtail resolver to new dir
